feat: 댓글 삭제 기능

+++ src/main/java/com/project/myblog/repository/ReplyRepository.java
@@ -3,5 +3,5 @@ package com.project.myblog.repository;
 import com.project.myblog.model.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReplyRepository extends JpaRepository<Reply,Integer> {
+public interface ReplyRepository extends JpaRepository<Reply,Long> {
 }
